requirements.txt
requests==2.21.0
scapy==2.4.3
Faker==2.0.3
SocksiPy_branch==1.01
termcolor==1.1.0
